fix(runtime-pi): move MCP utility modules out of extensions/
`mcp-direct.ts` and `mcp-provider-resolver.ts` are utility modules
imported by name from `entrypoint.ts` — not Pi extension factories.
But sitting under `runtime-pi/extensions/`, they were picked up by
the auto-discovery loop in `loadExtensionsFromDir()`, which expects
each .ts file to have a default function export. Result: every Pi
container run logged two spurious "default export is not a function
(got undefined)" errors at startup.

Move them to `runtime-pi/mcp/` to match their semantic purpose
(MCP-related helpers, not extensions) and stop the auto-discovery
from finding them. Update imports in entrypoint, both tests, and
the runtime-pi Dockerfile (COPY runtime-pi/mcp/ → ./mcp/).
